I originally made my deck tests in CasinoTest, fixed it; passed double in BlackJack so the program can be started from the Casino main
