+++ src/main/java/io/zipcoder/casino/BlackJack.java
@@ -1,7 +1,7 @@
 package io.zipcoder.casino;
 
 public class BlackJack extends CardGame{
-    Player dealer = new CardPlayer("Computer");
+    Player dealer = new CardPlayer("Computer", 100.00);
     public void startGame(Player player){
 
     }
